requirements.txt
numpy
pandas
sklearn
gensim
nltk
newspaper3k
matplotlib
seaborn
geopy
spacy
geojson